refactor(types): DRY up MergeRow, dedup ResolvedRow, and clean client type imports
