+++ apps/vscode-extension/resolver.ts
@@ -18,7 +18,7 @@ import { parseNotebook, serializeNotebook, renumberExecutionCounts } from '../..
 import { buildResolvedNotebookFromRows, type PreferredSide } from '../../packages/core/src/semanticResolution';
 import { WebConflictPanel } from './web/WebConflictPanel';
 import { UnifiedConflict, UnifiedResolution } from '../../packages/web/server/src/webTypes';
-import { NotebookSemanticConflict, Notebook } from '../../packages/core/src/types';
+import { NotebookSemanticConflict, Notebook, type ResolvedRow } from '../../packages/core/src/types';
 import * as gitIntegration from './gitIntegration';
 import { getSettings } from './settings';
 import * as logger from '../../packages/core/src/logger';
@@ -38,7 +38,7 @@ export const onDidResolveConflict = new vscode.EventEmitter<vscode.Uri>();
 export interface ResolvedConflictDetails {
     uri: vscode.Uri;
     resolvedNotebook?: Notebook;
-    resolvedRows?: import('../../packages/web/server/src/webTypes').ResolvedRow[];
+    resolvedRows?: ResolvedRow[];
     markAsResolved: boolean;
     renumberExecutionCounts: boolean;
     fileDeleted?: boolean;
@@ -611,7 +611,7 @@ export class NotebookConflictResolver {
     private async applySemanticResolutionsFromRows(
         uri: vscode.Uri,
         semanticConflict: NotebookSemanticConflict,
-        resolvedRows: import('../../packages/web/server/src/webTypes').ResolvedRow[],
+        resolvedRows: ResolvedRow[],
         markAsResolved: boolean,
         shouldRenumber: boolean,
         autoResolveResult?: AutoResolveResult,
